AgentGeneric

Description

few agents for generic games as defined in GenericGame

Documentation

class Agent2 a where

agent for Game2 games

Associated Types

type AgentParams a :: *

Methods

mkAgent :: AgentParams a -> ThrLocIO a

make new agent out of thin air using supplied neural network for evaluation

applyAgent :: (Game2 g, Repr (GameRepr g)) => a -> g -> Player2 -> ThrLocIO g

apply agent decision generating new game state playing as specified player.

agentName :: a -> String

get agent name

class Agent2Eval a where

Methods

evaluateGame :: (Game2 g, Repr (GameRepr g)) => a -> Player2 -> g -> Double

data AgEvalPlus a1 a2

summing agent, adds evaluations of any players it holds

Constructors

AgEvalPlus a1 a2

data AgEvalConst

constant player, everything is the same

Constructors

AgEvalConst Double

data AgentRandom

agent that picks a random move from all available moves

Constructors

AgentRandom GenIO

data AgentRandomSkew agEval

agent that picks a weighted random move from all available moves. weights are provided by game evaluator agent and ammended by bias

Constructors

AgentRandomSkew Double agEval GenIO

data AgentSimple

agent that picks a random move from all available *best* moves, where move fitness is determined by neural network evaluation

Constructors

AgentSimple TNetwork GenIO

data AgentMCTS

agent based on monte carlo tree search: evaluate moves counting how many times following that move and playing randomly till the end yields victory.

Constructors

AgentMCTS Int AgentRandom GenIO

data AgentParMCTS ag

agent based on monte carlo tree search: evaluate moves counting how many times following that move and playing randomly till the end yields victory.

Constructors

AgentParMCTS Double Int ag GenIO

data AgentGameTree

the usual alpha-beta etc. based agent.

Constructors

AgentGameTree TNetwork Int

data GameStateGen g

datatype useful for various game-tree-search based agents

Constructors

GameStateGen

Fields

gsgGame :: g

gsgEvalSelf :: GameStateGen g -> Int

gsgPlayerBase :: Player2

player that initialized tree search

gsgPlayerNow :: Player2

player which makes the move now

class HasGen a where

helper class for datatypes containing GenIO inside

Methods

gen :: a -> GenIO

class HasTNet a where

Methods

tnet :: a -> TNetwork

data AgentTrace a

Constructors

AgentTrace IOAct a

data IOAct

Constructors

IOAct (forall g. IO g -> IO g)

mkTimed :: (Agent2 a, AgentParams a ~ (IOAct, arg)) => String -> arg -> ThrLocIO a

timed :: Show t => t -> IO b -> ThrLocIO b

showTn :: TNetwork -> String

mkAgentSimpleFile :: FilePath -> ThrLocIO AgentSimple

pickList :: GenIO -> [a] -> IO a

pickListWeighted :: GenIO -> [(Double, a)] -> IO a

evalGameTNetwork :: (Game2 g, Repr (GameRepr g)) => TNetwork -> g -> Double

data GameDriverCallback g

data structure with callbacks for any occasion

Constructors

GameDriverCallback

Fields

gameFinished :: g -> IO ()

game finished possibly with a draw or we run out of time

gameTurn :: g -> Player2 -> IO Bool

new turn, player which takes move in it. return value True means continue with game

driverG2 :: (Game2 g, Repr (GameRepr g), Agent2 a1, Agent2 a2) => g -> a1 -> a2 -> GameDriverCallback g -> ThrLocIO g

a driver for Game2 games. passed appropriate callback structure drives the game from given state to the end.

sampleRandomGames

Arguments

:: (Repr (GameRepr g), Game2 g)

=> IO Bool
should we exit now?

-> Float
probability to call a callback on any generated game state

-> (g -> IO ())
callback

-> ThrLocIO ()

generate an infinite stream of random games, extracted via callback. exits when provided action returns False.

sampleRandomGamesCount

Arguments

:: (Repr (GameRepr g), Game2 g)

=> Int
callback call count

-> Float
probability to call a callback on any generated game state

-> (g -> IO ())
callback

-> ThrLocIO ()

generate an stream of random games, extracted via callback. exits after calling callback count times.

sampleRandomGameDepth

Arguments

:: (Repr (GameRepr g), Game2 g)

=> IO Bool
should we exit now?

-> (g -> Int -> IO ())
callback

-> ThrLocIO ()

generate an infinite stream depth of random games, extracted via callback. exits when provided action returns False.

sampleGameDepthCount

Arguments

:: (Repr (GameRepr g), Game2 g, Agent2 a1, Agent2 a2)

=> a1

-> a2

-> Int
callback call count

-> (g -> Int -> IO ())
callback

-> ThrLocIO ()

generate a stream of games between given agents, extracted via callback. exits after calling callback given number of times.

sampleGameDepth

Arguments

:: (Repr (GameRepr g), Game2 g, Agent2 a1, Agent2 a2)

=> a1

-> a2

-> IO Bool
should we exit now?

-> (g -> Int -> IO ())
callback

-> ThrLocIO ()

generate an infinite stream depth of games between given agents, extracted via callback. exits when provided action returns False.

while :: Monad m => m Bool -> m () -> m ()