BreakthroughGame

Documentation

type Position

Arguments

= (Int, Int)
(x,y), x=column goes from 0 to width-1, y=row goes from 0 to height-1

type BoardMap = Map Position Player2

data Breakthrough

Constructors

Breakthrough

Fields

board :: BoardMap

map from position to P1 or P2 piece.

boardSize :: (Int, Int)

width, height. height >= 4, width >= 2. each player will be given exactly 2*width pieces. there are width many columns and height many rows.

winningP1 :: Set Position

fixed set of winning positions for P1

winningP2 :: Set Position

fixed set of winning positions for P2

countP1 :: !Int

how many pieces P1 have

countP2 :: !Int

how many pieces P2 have

class Game2 a => Game2Features a where

Associated Types

type Feature a :: *

type FeatureRepr a :: *

Methods

encodeAllFeatures :: a -> FeatureRepr a

encodeFeature :: a -> Feature a -> FeatureRepr a

countFeatures :: a -> Int

data BrFeature

Constructors

MaxPos

MinPos

Centroid

Count

illegalPos :: Position -> Breakthrough -> Bool

legalPos :: Position -> Breakthrough -> Bool

row :: Int -> Int -> [Position]

get all positions within a specified row.

column :: Int -> Int -> [Position]

get all positions within a specified column.

allPos :: (Int, Int) -> [Position]

get all possible positions

getAllSet :: Player2 -> BoardMap -> Set Position

get all positions of pieces of specified player on board (as a set)

getAll :: Player2 -> BoardMap -> [Position]

get all positions of pieces of specified player on board (as a list)

count :: Player2 -> BoardMap -> Int

count number of pieces of specified player on board