Constraints

Documentation

newtype WBoard

Constructors

WBoard Board

newtype BBoard

Constructors

BBoard Board

data Constraint

Constructors

CBetter Board Board

CBetterAll Board [Board]

checkConstraint :: (Board -> Double) -> Constraint -> Bool

check if constraint is held under given evaluation function

scoreConstraints :: (Board -> Double) -> [Constraint] -> Double

score constraints, returns percentage true

generateConstraintsSimple :: Board -> Board -> [Constraint]

simple way to generate constraints, we dont do any recursion: just check immidiate moves

good'moves :: [(Board, Board)]

good moves

singleNeuronRandomSearch :: PrintfArg t => ((([[[Double]]], [[Double]]), Double, IO ()) -> IO a) -> Double -> t -> IO (([[[Double]]], [[Double]]), Double)

load DBN, apply single neuron to DBN output

benchmark random single-neuron networks

singleNeuronLocalSearch :: ((([[[Double]]], [[Double]]), Double, IO ()) -> IO ()) -> IORef (([[[Double]]], [[Double]]), Double) -> Double -> Double -> Int -> IO (([[[Double]]], [[Double]]), Double)

readGame :: Board -> String -> [Board]

(<>) :: a -> b -> (a, b)

movesFromGame :: (Color, [Board]) -> [(Board, Board)]

games'all :: [(Color, [Board])]

game1 :: (Color, [Board])

game5 :: (Color, [Board])

game4 :: (Color, [Board])

game3 :: (Color, [Board])

game2 :: (Color, [Board])