ConstraintsGeneric

Documentation

data Constraint g

Constructors

CBetter !g !g
first state is better than the other. assumes this is P1 turn.

CBetterAll !g ![g]
first state is better then ALL the others. assumes this is P1 turn.

type SingleNeuron = ([[[Double]]], [[Double]])

checkConstraint :: (g -> Double) -> Constraint g -> Bool

check if constraint is held under given evaluation function

scoreConstraints :: (g -> Double) -> [Constraint g] -> Double

score constraints, returns percentage true

generateConstraintsSimple :: (Game2 g, Eq g) => g -> g -> [Constraint g]

generate constraints using base game state and a best choice following that state

generateConstraintsSimpleAll :: (Game2 g, Eq g) => g -> g -> [Constraint g]

generate constraints using base game state and a best choice following that state

generateConstraintsMCTS :: (Repr (GameRepr g), Game2 g) => Int -> g -> ThrLocIO (g, g)

generate constraints with generateConstraintsSimple and AgentMCTS

generateConstraintsMCTS' :: (Repr (GameRepr g), Game2 g) => AgentMCTS -> g -> ThrLocIO (g, g)

generate constraints with generateConstraintsSimple and AgentMCTS

generateConstraintsGameplay :: Game2 g => Player2 -> [g] -> [(g, g)]

singleNeuronRandomReprSearch

Arguments

:: ((SingleNeuron, Double, IO ()) -> IO a)
callback called for new best score

-> Double
target value for score

-> Int
thread number

-> [Constraint (Vector Double)]
a list of good moves: [(first state, second state)]

-> ThrLocIO (SingleNeuron, Double)
(best'neuron, best'score) pair

singleNeuronRandomSearch

Arguments

:: (Eq g, Repr (GameRepr g), Game2 g)

=> ((SingleNeuron, Double, IO ()) -> IO a)
callback called for new best score

-> Double
target value for score

-> Int
thread number

-> FilePath
file with NN to read

-> [(g, g)]
a list of good moves: [(first state, second state)]

-> ThrLocIO (SingleNeuron, Double)
(best'neuron, best'score) pair

singleNeuronLocalReprSearch

Arguments

:: ((SingleNeuron, Double, IO ()) -> IO a)
callback called for new best neuron

-> IORef (SingleNeuron, Double)
best neuron IORef. **WARNING**: make sure to update this IORef with values passed to callback.

-> Double
base local search range

-> Double
target value

-> Int
thread num (affects effective search range)

-> [Constraint (Vector Double)]
a list of constraints

-> ThrLocIO (SingleNeuron, Double)
(best'neuron, best'score) pair

simplified and possibly faster version of singleNeuronLocalSearch. constraints need to have preapplied game and toRepr functions as well as response from DBN.

singleNeuronLocalSearch

Arguments

:: (Eq g, Repr (GameRepr g), Game2 g)

=> ((SingleNeuron, Double, IO ()) -> IO ())
callback called for new best neuron

-> IORef (SingleNeuron, Double)
best neuron IORef. **WARNING**: make sure to update this IORef with values passed to callback.

-> Double
base local search range

-> Double
target value

-> Int
thread num (affects effective search range)

-> FilePath
filename with NN

-> [(g, g)]
a list of good moves: [(first state, second state)]

-> ThrLocIO (SingleNeuron, Double)
(best'neuron, best'score) pair

neginf :: Double

scoreNeuron :: (Repr (GameRepr g), Game2 g) => SingleNeuron -> TNetwork -> [Constraint g] -> Double

evalNeuronGame :: (Repr (GameRepr g), Game2 g) => SingleNeuron -> TNetwork -> g -> Double

fixme: merge with AgentSimpleLL code