GenericGame

Description

generic interface for games

Documentation

data Player2

Constructors

P1
player 1

P2
player 2

nextPlayer :: Player2 -> Player2

nextPlayer P1 = P2 | nextPlayer P2 = P1

class GameTxtRender g where

Methods

prettyPrintGame :: g -> String

class Game2 a where

class for 2 player games

Associated Types

type MoveDesc a :: *

type for move descriptions

type GameRepr a :: *

type for game state representations

type GameParams a :: *

type for game parameters (board size etc.)

Methods

gameName :: a -> String

return game name

invertGame :: a -> a

reflect the game so that pieces etc. for P1 and P2 are inverted. throw error if this isn't possible.

freshGame :: GameParams a -> a

initiate new game

freshGameDefaultParams :: a

initiate new game using default parameters

movesDesc :: a -> Player2 -> [(MoveDesc a, a)]

get valid moves for a specified player as a list of new states and a move descriptor. if no moves can be made should return empty list. a passed move is allowed by including original state in output list and appropriate moveDesc.

moves :: a -> Player2 -> [a]

same as movesDesc but without descriptors. default implementation could be overriden for efficiency.

applyMove :: a -> MoveDesc a -> Maybe a

apply move description to get a new valid state. for invalid moveDesc Nothing shall be returned.

winner :: a -> Maybe Player2

returns a winner if the game is finished. can return either of players if both have the winning condition

serializeGame :: a -> ByteString

serialize game to bytestring. default implementation uses serializeRepr and toRepr

deserializeGame :: GameParams a -> ByteString -> a

deserialize game from bytestring. default implementation uses deserializeRepr and fromRepr

toRepr :: a -> GameRepr a

for neural network training. does not have to preserve game parameters.

toReprNN :: a -> Vector Double

for neural network training. does not have to preserve game parameters. provided for possibility of more efficient implementation.

fromRepr :: GameParams a -> GameRepr a -> a

load from game representation and game parameters

class Repr gameRepr where

representing games states

Methods

serializeRepr :: gameRepr -> ByteString

serialize representation to bytestring

deserializeRepr :: ByteString -> gameRepr

deserialize representation from bytestring

reprToNN :: gameRepr -> Vector Double

transform representation into vector to be fed into NN (for NN evaluation of game state)

class OneZero a where

Methods

zero, one :: a