MinimalNN

Description

minimal NN implementation based on hmatrix

Documentation

data TNetwork

Constructors

TNetwork

Fields

weights :: ![Matrix Double]

biases :: ![Vector Double]

computeTNetworkSigmoid :: TNetwork -> Vector Double -> Vector Double

computeTNetworkSigmoidSteps :: Int -> TNetwork -> Vector Double -> Vector Double

mkTNetwork :: [[[Double]]] -> [[Double]] -> TNetwork

appendNetwork

Arguments

:: TNetwork
first network, will receive input first

-> TNetwork
second network, will be fed output of first network

-> TNetwork
resulting network

append one network after another. input/output sizes must work out (which isn't checked until later)

lastLayerTN :: TNetwork -> [(Double, [Double])]

return last layer in nn, assuming it is one neuron big

sigmoid :: Floating a => a -> a