Tournament

Documentation

class AgentMutable a where

Methods

mutate :: GenIO -> Double -> a -> IO a

data EvolutionParameters

Constructors

EvolutionParameters

Fields

populationSize :: Int

championSquad :: Int

deathProbability :: Double

mutationForce :: Double

stepsToGo :: Int

plateuDetector :: ()

generator :: Hidden GenIO

data Population a

Constructors

Population

Fields

members :: [(Score, a)]

championScore :: Score

type Score = Double

getMemberScore :: (Score, a) -> Score

data TournamentRules

Constructors

TournamentRules

Fields

versusGames :: Int

maximumMoves :: Int

versusRandomCount :: Int

playVersus :: (Agent a, Agent b) => String -> Color -> Int -> Int -> Board -> a -> b -> IO Board

run a single game for maximum given number of turns. takes the color of current player.

playVersusTrace :: (Agent a, Agent b) => String -> Color -> Int -> Int -> Board -> a -> b -> IO (Maybe Color, [Board])

run a single game for maximum given number of turns. takes the color of current player. returns whole history of moves.

simpleGame :: Int -> IO Board

simpleGameTrace :: Int -> IO (Maybe Color, [Board])

tournament :: (Eq a, Agent a) => [a] -> IO [(Score, a)]

run a single round of tournament. players play versus each one.

playVersusRandom :: Agent a => a -> IO Score

play a game versus AgentRandom a number of times. receive score based on a number of games won.

cleanPopulation :: EvolutionParameters -> Population a -> IO (Population a)

remove members of population that die. RIP.

mutatePopulation :: AgentMutable a => EvolutionParameters -> Population a -> IO (Population a)

mutate a random member of population until we reach populationSize.

tournamentPopulation :: (Eq a, Agent a, Ord a) => Population a -> IO (Population a)

apply tournament evalution for population

evolutionStep :: (Ord a, Agent a, AgentMutable a) => EvolutionParameters -> Population a -> IO (Population a)

run a single evolution step

fillIOGaps :: EvolutionParameters -> IO EvolutionParameters

initialize parts that need IO monad

mkInitialPopulation :: (Agent a, AgentMutable a) => EvolutionParameters -> IO (Population a)

create initial population of a given type

evolution :: (Ord a, Eq a, Show a, Agent a, AgentMutable a) => Maybe (Population a) -> EvolutionParameters -> IO (Population a)

run evolution algorithm. without given population will start with fresh one.

decSteps :: EvolutionParameters -> EvolutionParameters

prettyReportPopulation :: Show a => Population a -> IO ()